requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
cryptography==43.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
serverless-wsgi==3.0.2
